feat: add get drinks api
