+++ sulsul-domain/src/test/kotlin/will/of/d/sulsul/alcohol/drinkingLimit/service/DrinkingLimitServiceTest.kt
@@ -7,9 +7,9 @@ import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import will.of.d.sulsul.SharedContext
-import will.of.d.sulsul.alcohol.Drink
 import will.of.d.sulsul.alcohol.drinkingLimit.domain.DrinkingLimit
 import will.of.d.sulsul.alcohol.drinkingLimit.repository.DrinkingLimitRepository
+import will.of.d.sulsul.drink.domain.Drink
 
 class DrinkingLimitServiceTest(
     private val drinkingLimitService: DrinkingLimitService,
